requirements.txt
alembic==1.13.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
dominate==2.9.1
email_validator==2.1.1
Flask==3.0.3
Flask-Bootstrap==3.3.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Font-Awesome-Flask==0.1.4
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
numpy==1.26.4
pandas==2.2.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.2
tzdata==2024.1
visitor==0.1.3
Werkzeug==3.0.3
WTForms==3.1.2
